refactor: main package 생성
