+++ RemindAlarm/app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.RemindAlarm"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.ReminderActivity"
+            android:name=".presentation.ui.main.ReminderActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
