+++ RemindAlarm/app/src/main/res/layout/activity_reminder.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.ReminderActivity">
+    tools:context=".presentation.ui.main.ReminderActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host"
