+++ RemindAlarm/app/src/main/res/layout/fragment_reminder_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".presentation.ui.ReminderMainFragment">
+    tools:context=".presentation.ui.main.ReminderMainFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layoutAddReminder"
